+++ src/components/button/Button.js
@@ -1,12 +1,14 @@
-import styled, { css } from "styled-components/native";
+import { ButtonContainer, ButtonText } from "./Button.styles";
 import { moderateScale } from "react-native-size-matters";
 import theme from "../../styles/theme";
 
-export const Button = ({ color, size, onPress, title }) => (
+const Button = ({ color, size, onPress, title }) => (
   <ButtonContainer onPress={onPress} color={color} size={getButtonSize(size)}>
     <ButtonText color={color} size={getFontSize(size)}>{title}</ButtonText>
   </ButtonContainer>
 );
+
+export default Button;
 
 const getButtonSize = (size) => {
   const variantButtonSize = {
@@ -25,48 +27,3 @@ const getFontSize = (size) => {
   }
   return variantFontSize[size] || theme.fontSizes.medium
 }
-
-const buttonPrimaryStyles = css`
-  background-color: ${({ theme }) => theme.colors.primary};
-  border-color: ${({ theme }) => theme.colors.primary};
-  border-width: ${moderateScale(1)}px;
-`;
-
-const buttonTextPrimaryStyles = css`
-  color: ${({ theme }) => theme.colors.accent};
-`;
-
-const buttonSecondaryStyles = css`
-  background-color: ${({ theme }) => theme.colors.accent};
-  border-color: ${({ theme }) => theme.colors.primary};
-  border-width: ${moderateScale(1)}px;
-`;
-
-const buttonTextSecondaryStyles = css`
-  color: ${({ theme }) => theme.colors.primary};
-`;
-
-const buttonStyles = {
-  primary: buttonPrimaryStyles,
-  secondary: buttonSecondaryStyles,
-}
-
-const buttonTextStyles = {
-  primary: buttonTextPrimaryStyles,
-  secondary: buttonTextSecondaryStyles,
-}
-
-const ButtonContainer = styled.TouchableOpacity`
-  margin-bottom: ${({ theme }) => theme.spacing.s}px;
-  justify-content: center;
-  padding: ${(props) => props.size}px;
-  border-radius: ${moderateScale(8)}px;
-  ${(props) => buttonStyles[props.color]};
-
-`;
-
-const ButtonText = styled.Text`
-  font-size: ${(props) => props.size}px;
-  text-align: center;
-  ${(props) => buttonTextStyles[props.color]};
-`;
