--- src/components/index.js
@@ -1,2 +1,0 @@
-export { Button } from './button/Button'
-export { Select } from './select/Select'
